server/requirements.txt
fastapi
uvicorn
PyJWT
pydantic